+++ src/main/java/com/webbanhang2/dao/ProductDaoImpl.java
@@ -275,8 +275,8 @@ public class ProductDaoImpl implements ProductDao {
             int productRoomId = p.getProductRoomId();
             String thumbnail = p.getThumbnail();
             String descipt = p.getDescription();
-            int quantity = p.getQuantity();
-            double price = p.getPrice();
+            long quantity = p.getQuantity();
+            long price = p.getPrice();
             String sql = "insert into product"
                     + "(product_code, product_name,product_categories_id,product_material_id,product_room_id,thumbnail,description,quantity,price) "
                     + "values(?,?,?,?,?,?,?,?,?)";
@@ -380,7 +380,7 @@ public class ProductDaoImpl implements ProductDao {
         public void setValues(PreparedStatement ps, int i) throws SQLException {
             Product p = productList.get(i);
             ps.setString(1, p.getProductId());
-            ps.setInt(2, p.getQuantity());
+            ps.setLong(2, p.getQuantity());
         }
 
         @Override
@@ -403,8 +403,8 @@ public class ProductDaoImpl implements ProductDao {
             product.setProductId(rs.getString("product_id"));
             product.setProductName(rs.getString("product_name"));
             product.setThumbnail(rs.getString("thumbnail"));
-            product.setPrice(rs.getInt("price"));
-            product.setQuantity(rs.getInt("quantity"));
+            product.setPrice(rs.getLong("price"));
+            product.setQuantity(rs.getLong("quantity"));
             return product;
         }
     }
@@ -425,8 +425,8 @@ public class ProductDaoImpl implements ProductDao {
             product.setProductOriginId(rs.getInt("product_origin_id"));
             product.setProductRoomId(rs.getInt("product_room_id"));
             product.setDescription(rs.getString("description"));
-            product.setQuantity(rs.getInt("quantity"));
-            product.setPrice(rs.getInt("price"));
+            product.setQuantity(rs.getLong("quantity"));
+            product.setPrice(rs.getLong("price"));
             product.setThumbnail(rs.getString("thumbnail"));
             return product;
         }
@@ -456,8 +456,8 @@ public class ProductDaoImpl implements ProductDao {
 //                    + "where`product_code` = '"
 //                    + product.getProductId() + "';";
 //            product.setImgList(jdbcTemplate.query(sql, new ImgNameMapper()));
-            product.setQuantity(rs.getInt("quantity"));
-            product.setPrice(rs.getInt("price"));
+            product.setQuantity(rs.getLong("quantity"));
+            product.setPrice(rs.getLong("price"));
             product.setProduct_top(rs.getBoolean("product_top"));
             return product;
         }
